+++ packages/ui5-application-writer/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sap-ux/ui5-application-writer
 
+## 0.12.6
+
+### Patch Changes
+
+-   Updated dependencies [5b5355c]
+    -   @sap-ux/ui5-config@0.13.1
+
+## 0.12.5
+
+### Patch Changes
+
+-   Updated dependencies [56409d0]
+    -   @sap-ux/ui5-config@0.13.0
+
+## 0.12.4
+
+### Patch Changes
+
+-   c18fc5c: chore(open-ux-tools) update devDependencies and change dependabot config
+-   Updated dependencies [c18fc5c]
+    -   @sap-ux/ui5-config@0.12.3
+
 ## 0.12.3
 
 ### Patch Changes